move page components into folders and rename to index.tsx
