+++ src/pages/Home/index.tsx
@@ -1,28 +1,28 @@
 import { FC, useCallback, useContext, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { toast } from "react-toastify";
-import AppStatusCard from "../components/Home/AppStatusCard/AppStatusCard";
-import BitcoinCard from "../components/Home/BitcoinCard/BitcoinCard";
-import ConnectionCard from "../components/Home/ConnectionCard/ConnectionCard";
-import HardwareCard from "../components/Home/HardwareCard/HardwareCard";
-import LightningCard from "../components/Home/LightningCard/LightningCard";
-import TransactionCard from "../components/Home/TransactionCard/TransactionCard";
-import TransactionDetailModal from "../components/Home/TransactionCard/TransactionDetailModal/TransactionDetailModal";
-import WalletCard from "../components/Home/WalletCard/WalletCard";
-import ListChannelModal from "../components/Shared/ListChannelModal/ListChannelModal";
-import LoadingSpinner from "../components/Shared/LoadingSpinner/LoadingSpinner";
-import OpenChannelModal from "../components/Shared/OpenChannelModal/OpenChannelModal";
-import ReceiveModal from "../components/Shared/ReceiveModal/ReceiveModal";
-import SendModal from "../components/Shared/SendModal/SendModal";
-import UnlockModal from "../components/Shared/UnlockModal/UnlockModal";
-import { useInterval } from "../hooks/use-interval";
-import useSSE from "../hooks/use-sse";
-import { AppStatus } from "../models/app-status";
-import { Transaction } from "../models/transaction.model";
-import { AppContext } from "../store/app-context";
-import { checkError } from "../util/checkError";
-import { instance } from "../util/interceptor";
-import { enableGutter } from "../util/util";
+import AppStatusCard from "../../components/Home/AppStatusCard/AppStatusCard";
+import BitcoinCard from "../../components/Home/BitcoinCard/BitcoinCard";
+import ConnectionCard from "../../components/Home/ConnectionCard/ConnectionCard";
+import HardwareCard from "../../components/Home/HardwareCard/HardwareCard";
+import LightningCard from "../../components/Home/LightningCard/LightningCard";
+import TransactionCard from "../../components/Home/TransactionCard/TransactionCard";
+import TransactionDetailModal from "../../components/Home/TransactionCard/TransactionDetailModal/TransactionDetailModal";
+import WalletCard from "../../components/Home/WalletCard/WalletCard";
+import ListChannelModal from "../../components/Shared/ListChannelModal/ListChannelModal";
+import LoadingSpinner from "../../components/Shared/LoadingSpinner/LoadingSpinner";
+import OpenChannelModal from "../../components/Shared/OpenChannelModal/OpenChannelModal";
+import ReceiveModal from "../../components/Shared/ReceiveModal/ReceiveModal";
+import SendModal from "../../components/Shared/SendModal/SendModal";
+import UnlockModal from "../../components/Shared/UnlockModal/UnlockModal";
+import { useInterval } from "../../hooks/use-interval";
+import useSSE from "../../hooks/use-sse";
+import { AppStatus } from "../../models/app-status";
+import { Transaction } from "../../models/transaction.model";
+import { AppContext } from "../../store/app-context";
+import { checkError } from "../../util/checkError";
+import { instance } from "../../util/interceptor";
+import { enableGutter } from "../../util/util";
 
 const startupToastId = "startup-toast";
